Use BIGINT for stack amount - fixes SQL error

+++ database/migrations/CreateStacks.php
@@ -16,7 +16,7 @@ class CreateStacks extends Database
 
         $sql = "CREATE TABLE stacks (
             id INT(6) UNSIGNED AUTO_INCREMENT PRIMARY KEY,
-            amount INT(12) UNSIGNED NULL,
+            amount BIGINT(12) UNSIGNED NULL,
             player_id INT(6) UNSIGNED NOT NULL,
             table_id INT(6) UNSIGNED NOT NULL,
             FOREIGN KEY (player_id) REFERENCES players(id),
